Support XR and improve performance & quality
- Add support for XR platforms
- Improve clouds blending with opaque objects to match the clouds behavior in HDRP
- Remove the need to adjust camera far plane for clouds to match the clouds behavior in HDRP
- Optimize performance by allowing fewer primary steps (24) and lower clouds resolution (0.4)

+++ Assets/VolumetricClouds/VolumetricCloudsDefs.hlsl
@@ -2,12 +2,12 @@
 #define URP_VOLUMETRIC_CLOUDS_DEFINES_HLSL
 
 CBUFFER_START(UnityPerMaterial)
-half _Seed;
+float _Seed;
 half _NumPrimarySteps;
 half _NumLightSteps;
 half _MaxStepSize;
-half _HighestCloudAltitude;
-half _LowestCloudAltitude;
+float _HighestCloudAltitude;
+float _LowestCloudAltitude;
 half4 _ShapeNoiseOffset;
 half _VerticalShapeNoiseOffset;
 half4 _WindDirection;
